moved S.Client.BaseUrl to S.Utils.BaseUrl
