+++ servant/src/Servant.hs
@@ -7,7 +7,7 @@ module Servant (
   module Servant.Server,
   -- | For accessing servant APIs as API clients.
   module Servant.Client,
-  module Servant.Client.BaseUrl,
+  module Servant.Utils.BaseUrl,
   -- | For generating documentation for servant APIs.
   module Servant.Docs,
   -- | Helper module
@@ -16,7 +16,7 @@ module Servant (
 
 import Servant.API
 import Servant.Client
-import Servant.Client.BaseUrl
 import Servant.Docs
 import Servant.Server
+import Servant.Utils.BaseUrl
 import Servant.Utils.Text
